microservice_ml/requirements.txt
# File: requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
pydantic==2.11.7
anyio==4.9.0
starlette==0.46.2
paho-mqtt

# Paket ML/CV
dlib==19.24.1
face-recognition==1.3.0
face_recognition_models==0.3.0
opencv-python-headless
pytesseract==0.3.13
pillow==11.3.0
numpy==1.26.4
mediapipe
requests